Revert "Merge branch 'master' into unicode_gamma"
This reverts commit ba43621e456aacb42dbbf23d1f911f5a180d2790, reversing
changes made to d296c8c4fb1b52539921acd436bbcba0e3a65c62.

+++ src/tasks.h
@@ -37,7 +37,7 @@ extern int tasks_set_connection_option(task_queue, const char *,
 
 extern void new_input_task(task_queue, const char *);
 extern enum error enqueue_forked_task2(activation a, int f_index,
-			       double after_seconds, int vid);
+			       unsigned after_seconds, int vid);
 
 extern enum error enqueue_suspended_task(vm the_vm, void *data);
 extern enum error enqueue_yielded_task(vm the_vm, void *data);
